+++ packages/angular_devkit/build_webpack/builders.json
@@ -2,23 +2,23 @@
   "$schema": "../architect/src/builders-schema.json",
   "builders": {
     "build": {
-      "implementation": "./src/webpack",
-      "schema": "./src/webpack/schema.json",
+      "implementation": "./src/builders/webpack",
+      "schema": "./src/builders/webpack/schema.json",
       "description": "Build a webpack app."
     },
     "webpack": {
-      "implementation": "./src/webpack",
-      "schema": "./src/webpack/schema.json",
+      "implementation": "./src/builders/webpack",
+      "schema": "./src/builders/webpack/schema.json",
       "description": "Build a webpack app."
     },
     "dev-server": {
-      "implementation": "./src/webpack-dev-server",
-      "schema": "./src/webpack-dev-server/schema.json",
+      "implementation": "./src/builders/webpack-dev-server",
+      "schema": "./src/builders/webpack-dev-server/schema.json",
       "description": "Serve a webpack app."
     },
     "webpack-dev-server": {
-      "implementation": "./src/webpack-dev-server",
-      "schema": "./src/webpack-dev-server/schema.json",
+      "implementation": "./src/builders/webpack-dev-server",
+      "schema": "./src/builders/webpack-dev-server/schema.json",
       "description": "Serve a webpack app."
     }
   }
