+++ packages/angular_devkit/build_webpack/package.json
@@ -6,6 +6,15 @@
   "main": "src/index.js",
   "typings": "src/index.d.ts",
   "builders": "builders.json",
+  "exports": {
+    ".": {
+      "types": "./src/index.d.ts",
+      "default": "./src/index.js"
+    },
+    "./package.json": "./package.json",
+    "./*": "./*.js",
+    "./*.js": "./*.js"
+  },
   "dependencies": {
     "@angular-devkit/architect": "0.0.0-EXPERIMENTAL-PLACEHOLDER",
     "rxjs": "7.8.1"
