+++ packages/angular_devkit/build_webpack/src/builders/webpack-dev-server/index_spec.ts
@@ -41,7 +41,7 @@ describe('Dev Server Builder', () => {
   }
 
   beforeEach(async () => {
-    const ngJsonPath = path.join(__dirname, '../../test/basic-app/angular.json');
+    const ngJsonPath = path.join(__dirname, '../../../test/basic-app/angular.json');
     const workspaceRoot = path.dirname(require.resolve(ngJsonPath));
     await createArchitect(workspaceRoot);
   });
